Added "createSimplePoolV2" on TerminusFacet
Also made it so that if `requiredPayment` is set to 0 on a Terminus
contract, we will not even initiate the ERC20 `transferFrom`. This makes
it so that a Terminus contract in which the controller does not have to
pay to create a pool does not require setting a payment token.

Bumped dao version to `0.0.8`.

Added tests for `createSimplePoolV2` functionality.